requirements/dev.txt
-r base.txt
pytest>=8.0,<9.0
pytest-django>=4.9,<5.0
pytest-cov>=6.0,<7.0
ruff>=0.8,<1.0
